requirements.txt
opensearch-dsl>=2.1.0
opensearch-py>=2.8.0
PyYAML>=6.0.2